// src/modules/misiones/types/idMissions/id-mission-list.enum.ts
export enum IdMissionsList {
    MISION_DE_CAZA_1 = 1,
    MISION_DE_CAZA_2 = 2,
    MISION_DE_CAZA_3 = 3,
    MISION_DE_CAZA_4 = 4,
    MISION_DE_CAZA_5 = 5,
    MISION_DE_CAZA_6 = 6,
    MISION_DE_CAZA_7 = 7,
    MISION_DE_CAZA_8 = 8,
    MISION_DE_CAZA_9 = 9,
    MISION_DE_CAZA_10 = 10,
    MISION_DE_CAZA_11 = 11,
    MISION_DE_CAZA_12 = 12,
    MISION_DE_CAZA_13 = 13,
    MISION_DE_CAZA_14 = 14,
    MISION_DE_CAZA_15 = 15,
    MISION_DE_CAZA_16 = 16,
    MISION_DE_CAZA_17 = 17,
    MISION_DE_CAZA_18 = 18,
    MISION_DE_CAZA_19 = 19,
    MISION_DE_CAZA_20 = 20,
    MISION_DE_CAZA_21 = 21,
    MISION_DE_CAZA_22 = 22,
    MISION_DE_CAZA_23 = 23,
    MISION_DE_CAZA_24 = 24,
    MISION_DE_CAZA_25 = 25,
    MISION_DE_CAZA_26 = 26,
    MISION_DE_CAZA_27 = 27,
    MISION_DE_CAZA_28 = 28,
    MISION_DE_CAZA_29 = 29,
    MISION_DE_CAZA_30 = 30,
    MISION_DE_CAZA_31 = 31,
    MISION_DE_CAZA_32 = 32,
    MISION_DE_CAZA_33 = 33,
    MISION_DE_CAZA_34 = 34,
    MISION_DE_CAZA_35 = 35,


    UN_NUEVO_COMIENZO = 501,
    CONOCETE_A_TI_MISMO = 502,
    UN_FAVOR_DEL_HERRERO = 503,
    NACE_UN_ESPECIALISTA = 504,
    AROMAS_DE_ESPERANZA = 505,
    OTRO_FAVOR_MAS = 506,
    MISION_IMPOSIBLE = 507,
    LA_VENGANZA_DE_LOS_LOBOS_ALFA = 508,
    ECO_DE_LA_CAIDA_SOMBRIA = 509,
    SOMBRAS_DEL_EID_BLANCO = 510,
    EL_DESPERTAR_DE_SUNG_MAHI = 511,
    EL_SECRETO_DE_LAS_ROCAS_NETIM = 512,
    LA_PRIMERA_SEÑAL = 513,
    EN_BUSCA_DE_EXPLICACIONES = 514,
    LA_AMBICION_SALVAJE = 515,
    LA_SEGUNDA_SENAL = 516,
    ARENAS_MOVIDIZAS = 517,
    RECETARIO_I = 518,
    RECETARIO_II = 519,
    LA_TERCERA_SENAL = 520,

    INGREDIENTE_PARA_MEDICINA = 1000,
    EL_PIN_PERDIDO = 1001,
    LA_CAIDA_DE_EUN_JUNG = 1002,
    PRUEBA_DEL_JINETE_I_EL_SENOR_MONO = 1003,
    MI_JUNG_DE_LOS_EIDS_BLANCOS = 1004,
    PREPARATIVOS_DE_UN_CABALLERO = 1005,
    UN_ATUENDO_LEGENDARIO_I = 1006,
    UN_ATUENDO_LEGENDARIO_II = 1007,
    UN_ATUENDO_LEGENDARIO_III = 1008,
    UN_ATUENDO_LEGENDARIO_IV_FINAL = 1009,
    CONOCE_A_TU_ENEMIGO = 1010,
    SUMINISTROS_PERDIDOS = 1011,
}